# generated/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ServeMate API",
    "version": "1.0.0",
    "description": "Auto-generated OpenAPI from DTO/Zod schemas"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development server"
    }
  ],
  "paths": {
    "/api/auth/login": {
      "post": {
        "summary": "login",
        "operationId": "AuthenticationController_login",
        "tags": [
          "Authentication"
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  }
                },
                "required": [
                  "email",
                  "password"
                ]
              }
            }
          }
        }
      }
    },
...
    "/api/reservations/:id/tables": {
      "patch": {
        "summary": "updateTables",
        "operationId": "ReservationsController_updateTables",
        "tags": [
          "Reservations"
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "tables": {
                    "type": "array",
                    "items": {
                      "type": "number"
                    }
                  }
                },
                "required": [
                  "tables"
                ]
              }
            }
          }
        }
      }
    },
...
    "/api/tables": {
      "get": {
        "summary": "findAll",
        "operationId": "TablesController_findAll",
        "tags": [
          "Tables"
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          }
        },
        "parameters": [
          {
            "name": "value",
            "in": "query",
            "required": false,
            "schema": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "number"
                },
                "tableNumber": {
                  "type": "number"
                },
                "minCapacity": {
                  "type": "number"
                },
                "maxCapacity": {
                  "type": "number"
                },
                "status": {
                  "type": "string"
                },
                "isOccupied": {
                  "type": "boolean"
                },
                "page": {
                  "type": "number"
                },
                "pageSize": {
                  "type": "number"
                },
                "sortBy": {
                  "type": "string"
                },
                "sortOrder": {
                  "type": "string"
                }
              },
              "required": [
                "status",
                "isOccupied"
              ]
            }
          }
        ]
      },
      "post": {
        "summary": "create",
        "operationId": "TablesController_create",
        "tags": [
          "Tables"
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "tableNumber": {
                    "type": "number"
                  },
                  "capacity": {
                    "type": "number"
                  },
                  "additionalCapacity": {
                    "type": "number"
                  },
                  "status": {
                    "type": "string"
                  }
                },
                "required": [
                  "tableNumber",
                  "capacity"
                ]
              }
            }
          }
        }
      }
    },
    "/api/tables/assign": {
      "post": {
        "summary": "assign",
        "operationId": "TablesController_assign",
        "tags": [
          "Tables"
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "serverId": {
                    "type": "number"
                  },
                  "isPrimary": {
                    "type": "boolean"
                  },
                  "assignedTables": {
                    "type": "array",
                    "items": {
                      "type": "number"
                    }
                  }
                },
                "required": [
                  "serverId",
                  "assignedTables"
                ]
              }
            }
          }
        }
      }
    },
    "/api/tables/:id": {
      "get": {
        "summary": "findOne",
        "operationId": "TablesController_findOne",
        "tags": [
          "Tables"
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          }
        },
        "parameters": [
          {
            "name": "value",
            "in": "path",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "number"
                }
              },
              "required": [
                "id"
              ]
            }
          }
        ]
      },
      "put": {
        "summary": "update",
        "operationId": "TablesController_update",
        "tags": [
          "Tables"
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "tableNumber": {
                    "type": "number"
                  },
                  "capacity": {
                    "type": "number"
                  },
                  "additionalCapacity": {
                    "type": "number"
                  },
                  "status": {
                    "type": "string"
                  },
                  "isOccupied": {
                    "type": "boolean"
                  },
                  "guests": {
                    "type": "number"
                  }
                },
                "required": [
                  "isOccupied"
                ]
              }
            }
          }
        }
      },
      "delete": {
        "summary": "delete",
        "operationId": "TablesController_delete",
        "tags": [
          "Tables"
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          }
        },
        "parameters": [
          {
            "name": "value",
            "in": "path",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "number"
                }
              },
              "required": [
                "id"
              ]
            }
          }
        ]
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}